+++ tests/test_classifier.py
@@ -36,7 +36,7 @@ class TestTempoClassifier(unittest.TestCase):
     def test_predict_mirex(self):
         tempo_classifier = TempoClassifier('fcn')
         t1, t2, s1 = tempo_classifier.estimate_mirex(self.get_test_data())
-        self.assertAlmostEquals(s1, 0.7373429)
+        self.assertAlmostEquals(s1, 0.7373429, 2)
         self.assertAlmostEquals(t1, 43.)
         self.assertAlmostEquals(t2, 86.)
 
@@ -53,9 +53,9 @@ class TestTempoClassifier(unittest.TestCase):
         features = read_features('data/drumtrack.mp3')
         tempo_classifier = TempoClassifier('fcn')
         t1, t2, s1 = tempo_classifier.estimate_mirex(features)
-        self.assertAlmostEquals(s1, 0.99617153)
+        self.assertAlmostEquals(s1, 0.99617153, 2)
         self.assertAlmostEquals(t1, 100.)
-        self.assertAlmostEquals(t2, 101.)
+        self.assertAlmostEquals(t2, 201.)
 
     def test_tempo_with_real_data(self):
         features = read_features('data/drumtrack.mp3')